#4445 - Program deactivate action bug fix (#4750)
## Bug fix on common component shared between Institutions and Ministry

### Root Cause
- The common component has a condition to disable the `Deactivate`
button based on the institution user data from institution store. This
condition results in a runtime error when the component is loaded from a
ministry user view because of the absence of Institution store and so
the institution user data.

## Solution
- Validation of institution user data is applied conditionally when the
user type is `Institution` skipping the same for a ministry view.
Note: We generally avoid using the auth client type based logic in the
components `AuthService.shared.authClientType ===
ClientIdType.Institution`. But this component being very old and already
using similar condition in other places, the auth client type is used
here as quick solution.

### Outcome Ministry


![image](https://github.com/user-attachments/assets/b7dab337-8ddf-4fb2-97ae-8569fa550ce2)


### Outcome Institution(No changes introduced here by PR)


![image](https://github.com/user-attachments/assets/e9508d61-078a-48de-b017-500611f86ba5)