#4748 - 25/26 PY PT: Books and Supplies not displayed in assessed costs table (#4749)
Books and supplies was not showing in the assessed costs table because
`calculatedDataProgramRelatedCosts` was calculating to null (showing not
eligible in the NOA).
This was due to `programYearTotalPartTimeBooksAndSuppliesCost` not being
checked for being null.

- Added null check and moved all program year totals to the front of the
assessment flow:


![image](https://github.com/user-attachments/assets/556ed353-2bd2-468e-8781-2d255b8a043b)


### E2E Updates

**parttime-assessment-2025-2026-awards-amount-CSLP.** 
> Should determine CSLP award for an assessment which involves loan and
grants when the student does not have any loan balance and has remaining
need for loan.
> Should determine CSLP award for an assessment which involves loan and
grants when the student have loan balance and has remaining need for
loan.

Updated Expected Remaining Need +3000 to account for books and supplies
max cost

**parttime-assessment-2025-2026-program-related-costs.**
Updated tests to be around the program related costs total.