#4752 - FT Assessment tests - scholarship and bursaries (#4776)
- Fixed some linting issues on child care costs test file
- Added new scholarship and bursary tests:
    - $0 declared + $0 py total
    - under PY limit declared + $0 py total
    - over PY limit declared + $0 py total
    - over PY limit declared + $1000 py total
    - under PY limit declared + $1000 py total
- Added some variables to the assessment model to support the test cases

Scholarships & bursaries declared by the student will be considered
towards their contribution.
The student is allowed to have scholarships & bursaries NOT counted
towards contribution (exempt) up to a program year limit.
We need to track the amount of exemption remaining, the amount to exempt
from this application, and the amount to attribute to contribution in
this application.