#4740 - Parent update first effort - DB Migrations and First Workflows (#4788)
## DB Migrations

- Created new columns `is_able_to_report` and `full_name` in
`supporting_users`.
- Created a constraint to ensure the names will be unique for an
application (the same will be done in the UI later).

## Workers

- No workers changes were done as part of this first PR other than the
enum item `CreateIdentifiableSupportingUsers` that was added to allow
the execution of the E2E tests that use it to create the mocked workers.

## Assessment Gateway

- A new version of the assessment gateway was created from the most
recent `assessment-gateway` version to include the new process for the
parents' creation.
- Used the
[multi-instance](https://docs.camunda.io/docs/components/modeler/bpmn/multi-instance/)
as decided during the analysis ticket #4477.


![image](https://github.com/user-attachments/assets/ecb2809b-c659-4ba5-abac-540806f69f6e)

## Full-time assessment (temporary changes)

- Created a new version of the full-time assessment to allow the tests
to pass. No impacts to the applications are expected since the workflow
will be used only for E2E tests in this PR.

### Current state


![image](https://github.com/user-attachments/assets/33eb15a5-b6a5-4917-84d5-c4c23cac6b3d)

### Updated state


![image](https://github.com/user-attachments/assets/b28c4e3d-d865-4575-ad1d-825d527702f0)

## Consolidated data

- In preparation for the upcoming form.io changes, the new variable
mapping was added to load the new array of parents.

## Workflow E2E Tests

- Introduced a new constant `DEFAULT_ASSESSMENT_GATEWAY` in the 2025-26
program year tests to set the assessment gateway version to be used. The
gateway version is not directly coupled to a program year and can be set
differently for any E2E scenario as required. The constant is intended
to be used as the default unless there is a reason to specify something
different.

### New workers mock helper (multi-instance mock enablement)

- Created a new mock helper `createIdentifiableParentTaskMock` to mock
the data for the new task `create-identifiable-parent-task`.
- Changed the `WorkerMockedData` to allow the indication of which loop
the mocked data should be returned to, using the new property
`multiInstanceLoopCounter`.
- Updated the worker mock creation method (`createWorkersMockedData`) to
create an object that would indicate that the data belongs to a
multi-instance subprocess and allow data to be mocked to individual
instances. Please see below what it looks like.

```json
{
    "isMultiInstance": true,
    "instances": [
        {
            "result": {
                "createdSupportingUserId": 2000
            },
            "messageResult": [
                {
                    "name": "supporting-user-info-received",
                    "correlationKey": "2000",
                    "variables": {},
                    "timeToLive": {
                        "type": "SECONDS",
                        "value": 10,
                        "valueType": "TYPED_DURATION",
                        "unit": "s"
                    }
                }
            ]
        },
        {
            "result": {
                "createdSupportingUserId": 2001
            },
            "messageResult": [
                {
                    "name": "supporting-user-info-received",
                    "correlationKey": "2001",
                    "variables": {},
                    "timeToLive": {
                        "type": "SECONDS",
                        "value": 10,
                        "valueType": "TYPED_DURATION",
                        "unit": "s"
                    }
                }
            ]
        }
    ]
}
```
_Note:_ The intention of having the property `isMultiInstance` is to
make it clear and better to understand, while reading the code, the
differences between the processes.

For a comparison, the current worker non-multi-instance looks like the
below.

```json
{
    "result": {
          "someProperty": "somevalue"
      },
      "messageResult": [
          {
              "name": "messageName",
              "correlationKey": "2000",
              "variables": {},
              "timeToLive": {
                  "type": "SECONDS",
                  "value": 10,
                  "valueType": "TYPED_DURATION",
                  "unit": "s"
              }
          }
      ]
}
```

### Created new assertion items for multi-instance

- Adapted the method `expectToPassThroughServiceTasks` to execute the
assertion in a new class `WorkflowMultiInstanceProcess`, which contains
the information to extract the variable "passthrough" values to
determine if a flow was executed in the expected way.
- To allow the multi-instance to report the status for the individual
loop contexts, a collection output is set to allow the later inspection
by the method `expectToPassThroughServiceTasks`.


![image](https://github.com/user-attachments/assets/79c82c95-3824-4766-a0a0-8f58e49cfc5f)

### Non-direct changes

- The `createCheckIncomeRequestTaskMock` was not present in the workers
mocked data, and it was added to the 2025-26 E2E tests. The tests were
previously working because the worker mock `createIncomeRequestTaskMock`
was returning the `incomeVerificationCompleted`, which is not supposed
to be returned by the `createCheckIncomeRequestTaskMock`. The
`incomeVerificationCompleted` was not removed at this moment from the
`createCheckIncomeRequestTaskMock` because it would cause an impact in
approximately 64 changes in a few more files.

## Migration Rollback


![image](https://github.com/user-attachments/assets/be450130-34d1-4704-b3e4-69c00a8533d8)