#4740 - Parent update first effort - Application Submission (#4805)
## Student Application (sfaa2025-26-ft)

- Created a new table to allow one to two parents to be added, as shown
in the image below.
- Created a validation to prevent two parents from being provided with
the same name.
  - Executed the validation using a case-insensitive.
- Executed the `trim` to avoid white spaces from being considered as two
different strings.
- _Note:_ No extra validations were considered required at this moment
while comparing the two strings.
- At least one parent is required, and a maximum of two can be provided.


![image](https://github.com/user-attachments/assets/41b559a7-8645-4c64-8d51-9f27b8720f42)
Please note that the "(s)" was removed, and the image does not reflect
this specific content change.

- Configured the custom error message as below to allow other
validations, for instance, the required one, to continue to display its
own message.


![image](https://github.com/user-attachments/assets/82a0ded7-7ff9-412d-90d8-c63c8346c8c0)

## Workers

### create-income-request

- Adjusted the income request creation to return the
`canExecuteIncomeVerification` to indicate if the workflow needs to wait
for the income verification to be completed and returned by the CRA.
- The workflow was adjusted in a backward-compatible way to be used for
all program years.
- The highlight change below allows the same process of CRA record
creation to happen with the new possibility to no longer wait for a CRA
response.
_Note:_ Many changes are related to adding a meaningful name to the
workflow element for better understanding.


![image](https://github.com/user-attachments/assets/84f56d9a-45d5-4dcf-915c-8842d447be93)

## create-identifiable-supporting-users (new)

- Create a new worker to process the creation of the identifiable
parents (identifiable by the full name) with the potential to be used
for a partner in the future, even if the full name is not required.
- The full name was kept optional, keeping in mind that a partner may
not require it. For parents, the filter is always provided, which will
ensure a full name is extracted for the dynamic data, or an exception
will be thrown.

## assessment-gateway-v2

- Minor changes to the variables propagation outputs (not required), and
some inputs are explicitly set for better understanding only.
_Note:_ The input propagation is currently required to allow the mocked
workers' variables to flow inside the subprocess.

## Outside PR scope

- Change the workers' E2E tests to happen in sequence to prevent the
intermittent timeout issues.

## Upcoming PRs

- Merge `fulltime-assessment-v2-2025-2026` and
`fulltime-assessment-v2-2025-2026-v2`.
- Change the CRA scheduler.
- E2E tests