#4219 - Beta Mode for Full-Time (#4837)
- Added a new property `hasFulltimeAccess` to the student profile DTO
(also saved to Vue store) to allow the checks in the UI. _Note:_ This
variable was added aiming for the least amount of effort and changes
since it will be removed once the beta phase is done for full-time.
- Included the `hasFulltimeAccess` in the conditions to show the
`Full-time` while creating an application.
- Included the `hasFulltimeAccess` in the conditions to show the menu
for the `Overawards balance`. _Note:_ No restrictions were added to
display the page or to block the API from retrieving data for the
overawards.
- The existing `allowBetaUsersOnly` is still used to dictate when the
beta user configurations should be used (as stated in business ACs).

## E2E Tests
- StudentInstitutionsController(e2e)-getStudentProfile
- Should get the student profile with the hasFulltimeAccess as true user
when the student was added to the beta users table and
allowBetaUsersOnly is true.
- Should get the student profile with hasFulltimeAccess as false when
the student was added to the beta users table with a future date and
allowBetaUsersOnly is true.
- Should get the student profile with hasFulltimeAccess as false when
the student was not added to the beta users table and allowBetaUsersOnly
is true.
- Should get the student profile with hasFulltimeAccess as true when the
student was not added to the beta users table but allowBetaUsersOnly is
false.

- ApplicationStudentsController(e2e)-submitApplication
- Should submit a full-time application when the student is configured
as a beta user.
- Should throw a forbidden error when a full-time application is
submitted and the user is not a beta user.

_Note:_ `allowBetaUsersOnly` E2E helper is duplicated because it will
also be removed soon.

Please note the below comment introduced further E2E tests:
https://github.com/bcgov/SIMS/pull/4837#discussion_r2157632919