#4819 - Student submit parent information (#4862)
# Student report parent supporting details

## API

- [x] Student API to submit supporting user details for parents who are
not able to report.
- [x] API is validated for supporting user who is parent, who is not
able to report, supporting data not submitted already and dry run
validation.
- [x] API also validates if the `Application Status` is `In-Progress`
(For applications submitted/edited before COE) or `Application Edit
Status` is `Change in progress` (For applications edited post COE) to
allow the reporting of supporting data.
- [x] Created `SupportingUserControllerService` to abstract the logic
which can be potentially shared between the controllers. As of now it
has the method to validate the form dry run submission created in the
PR.

## Supporting user parent Form

- [x] Introduced a new hidden variable `isAbleToReport` based on which
the form allows personal information inputs(given names and last name)
and hides the `SIN` input and read only view of personal details

![image](https://github.com/user-attachments/assets/f68b19aa-deed-41b9-9368-dcb8ee8ff4fc)


## UI

- [x] Created component `SupportingUserForm` to display/edit support
user form and handle wizard navigation.


![image](https://github.com/user-attachments/assets/f4eba095-ed5f-4197-9d2d-2b34cdbb20b2)


![image](https://github.com/user-attachments/assets/dff9febd-9b94-477c-89ae-b24ea7cae133)


![image](https://github.com/user-attachments/assets/fdec23ab-78db-4396-bd07-0a65c1d09997)


![image](https://github.com/user-attachments/assets/137c8a76-cf4e-4935-a56c-ea6f4a6b083d)