#4895 - Appeal (Room and Board) - Camunda Updates (#4991)
**Calculation**
The actual calculations for this are done in line with where it exists
in the assessment so it can be properly included in the MSOL total when
approved.
Comparing the monthly amount provided in the appeal to the monthly
maximum to find the lower of the values before converting to a weekly
amount. After converting the monthly amount to weekly (divide by 4.3),
the value is then rounded (half-up) to align with the current MSOL being
only whole dollars.


**Additional MSOL Calculation Updates**
`calculatedDataStudentMSOLAllowance` was just a weekly value and was not
converted to a study period value until it was within
`calculatedDataTotalMSOLAllowance`.
This value is reported to institutions through IER12 and would have been
misleading as a weekly value.
Updated the BPMN to include `calculatedDataStudentTotalMSOLAllowance`
and updated the gateway `workflowData.calculatedData` to reflect the
change from "StudentMSOL..." to "StudentTotalMSOL..."

**Test Cases Added**
- Appeal:
   - No appeal
   - Appeal with less than max
   - Appeal with more than max
- Living Allowance:
    - SIA / 0 dep / BC
    - SIA / 0 dep / AB
    - SDH / BC
    - SP / 1 dep / BC
    - SP / 2 dep / BC
    - M / 0 dep / BC
    - M Unable to Report / 0 dep / BC

**BPMN Changes**
Lots of line changes showing on the BPMN to accommodate the flow
changes:

<img width="1047" height="707" alt="image"
src="https://github.com/user-attachments/assets/22d02bb6-55c3-4950-b0b0-db07502a6f53"
/>