#4941 - Approved exception does not repeat - New JSON Lib (#5010)
- Replaced the JSONPath with the
[jsonata](https://github.com/jsonata-js/jsonata) lib to allow the filter
and transformations that will be required once the new exceptions
structure is in place.
- Adapt the load-consolidate-data BPMN, and a few workflows that were
using the JSONPath expression.

## Important note about the expressions

Assuming the below JSON (array with a single item).
```json
{ 
  "courseDetails": [
      {
          "courseCode": "code",
          "courseName": "name"
      }
  ]
}
```
Using the expression below
```
courseDetails.{
  "courseCode": courseCode
}
```
will produce the result as an object, not an array.
```json
{
  "courseCode": "code"
}
```

To produce an array even when there is only one item, the expression
should be as below.
```
courseDetails[].{
  "courseCode": courseCode
}
```
to produce
```json
[
  {
    "courseCode": "code"
  }
]
```

## Manual tests
- Created a new 2025-26 full-time application including two parents (one
self-declare and the other doesn't) and also an exception.
- Created a change request to remove one of the parents.
- Create a full-time application including dependents.
- Created a 2025-26 part-time application.
- Created a 2024-25 application and an appeal to add dependents.