#4731 - Form-io loader (#5008)
## Form io loader refactor

- [x] Updated the `formio` and hence the `formio-container` component to
receive a state of `inital data loading is complete` by the new
parameter `is-data-ready` which indicates that all the data required to
load the form is ready now.
- [x] This indicator mentioned above `is-data-ready` will allow the form
to be created, and until the flag is set to `true` the the form creation
will be delayed.
- [x] During the form creation using `Formio.createForm` the initial
data is set during this process avoiding a form redraw.

## Loader

- [x] Form io component now uses the `skeleton` loader from Vuetify.
- [x] There is a slot provided for the consumer to use a custom skeleton
loader if that is required.

## Footer Buttons

- [x] Footer buttons are not displayed until the `is-data-ready` is true
as it is considered to be part of the form, even though they are
technically outside form.io.

## Updated forms

- [x] All the student portal forms which use initial data for the form
loading are updated.

## Skeleton loader


<img width="1226" height="422" alt="image"
src="https://github.com/user-attachments/assets/dac45ed3-f8e7-4ac2-82f3-c78303207f4d"
/>

<img width="1241" height="535" alt="image"
src="https://github.com/user-attachments/assets/d193d253-80ac-41d3-9864-725aec82e654"
/>