#4941 - Approved exception does not repeat - Hash Creation (#5030)
- Created a new worker method to handle the new exceptions strategy to
find and detect already approved exceptions.
- Create a new service (`ApplicationExceptionSearchService`) to search
for application exceptions, looking for a container with the suffix
"ApplicationException".
- Created a new service (`ApplicationExceptionHashService`) to generate
a unique hash that would change if the exception data, including its
files, changed. _Note:_ A file is considered "the same" when the
user-provided name (not the unique name with the guid) and its contents
are the same.
- Created a [type
guard](https://www.typescriptlang.org/docs/handbook/advanced-types.html)
(`isApplicationDataExceptionFile`) to detect a file object from the
dynamic data.
- Camunda workflows are adapted to receive new information to decide
which exception strategy to process. In the upcoming PRs, the new
strategy will be enabled for the full-time application alongside the new
exception containers in the application data.

## Data Sample

| Application Exception ID| Request ID | Exception Name | Exception
Description | Approval Exception Request ID | Exception Hash |

|-------------------------|-----|---------------------------------|-------------------------------------|------------------------------|-------------------------------------------------------------------|
| 79 | 151 | exceptionA | Exception A | 147 | 10e8f9d54... |
| 79 | 150 | exceptionB | Exception B - 1 | [NULL] | ff283cd20... |
| 78 | 149 | exceptionA | Exception A | 147 | 10e8f9d54... |
| 78 | 148 | exceptionB | Exception B - 1 | 146 | 99283cd20... |
| 77 | 147 | exceptionA | Exception A | [NULL] | 10e8f9d54... |
| 77 | 146 | exceptionB | Exception B - 1 | [NULL] | 99283cd20... |

Where 77, 78, and 79 each represent an application submission with the
same two exceptions.
- 77 was first submitted and had its exceptions approved.
- 78 was submitted and had the same exceptions detected and the same
hashes generated, hence the `Approval Exception Request ID` is populated
to indicate these were previously approved.
- 79 was submitted and had only one exception defined as already
approved. Please note the referenced approved exception`147` is the
expected one to be associated since it was the first ever approved
exception.

## E2E Tests

- Created the below main E2E tests to test two high-level scenarios. The
other E2Es on the file are the same as the existing worker.
- Should create and associate an application exception with three
exception requests with hashes when there are exceptions for the
application.
- Should create and associate an approved application exception with two
exception requests when all application exceptions part of the new
submission were previously approved.