#5160 - Container inconsistent red edit section highlighting (#5173)
Form.io API keys can be repeated if present in different containers,
which is now more common due to the new non-repeat-exceptions feature,
but it is also something naturally supported by Form.io.
While searching for the components to be highlighted as changed, the
first UI component matching the `key` was returned, which could
potentially point to any component, at any nested level, in the form.io
components tree.
Keeping in mind that the below JSON also does not represent or try to
mimic the **full** hierarchy of the components tree, the solution was to
get the nearest component from the reference component.

For reference, the JSON below represents the API output for the
application data changes, where the `key` `currentYearIncome`, is
present at two different levels.

```json
"changes": [
        {
            "key": "taxReturnIncome",
            "changeType": "updated"
        },
        {
            "key": "currentYearIncome",
            "changeType": "updated"
        },
        {
            "key": "currentYearIncomeApplicationException",
            "changeType": "updated",
            "changes": [
                {
                    "key": "currentYearIncome",
                    "changeType": "updated"
                }
            ]
        }
    ],
```