#4964 - Upgrade Packages - Camunda 8 SDK (#5222)
- `CAMUNDA_SECURE_CONNECTION` is no longer required and creates a
warning if present.
- The connection now needs to be explicitly defined using `grpc://` or
`grpcs://`, which will require a release instruction to be added to
change the current GitHub env `ZEEBE_ADDRESS`.
- The env variable used to define the Zeebe connection `ZEEBE_ADDRESS`
was replaced a few releases ago by `ZEEBE_GRPC_ADDRESS`, but both were
working, which is no longer true. With the removal of the
`CAMUNDA_SECURE_CONNECTION` and the requirement to prefix the address
with grpc/grpc, such configuration only worked when using the newer
`ZEEBE_GRPC_ADDRESS`.
- _Note:_ the GitHub variable and secret are kept as `zeebe-address` to
avoid further changes and further release manual instructions, also
considering the name `zeebe-address` is still meaningful.
- Local Camunda docker versions were updated to use the same version
from DEV/TEST/Staging.
-
[camunda-docker-compose-core.yml‎](https://github.com/bcgov/SIMS/pull/5222/files#diff-95b5b59a1b69248d37d3c3e37fccee1eb5eda083cdc7da7e9e3639b0c90d672d)
was update with the most recent
[docker-compose-core.yaml](https://docs.camunda.io/docs/next/self-managed/quickstart/developer-quickstart/docker-compose/).