#2913 - COE Records After Study End Date and Dashboard Updates - E2E Tests (#5257)
- Added COEs for the new offering end date limit and new offering
intensity filter.
- Should get the COE current summary only for applications with offering
end date till today when there is one COE for an offering with end date
as today, one COE with the offering end date as tomorrow, and one COE
with the offering end date in the past.
- Should get the COE current summary only for full-time applications
when there is one COE for a full-time and one for a part-time
application.