#5183 - New appeals UI appeal history (#5343)
## New "Appeal History" tab in the Student Appeals

<img width="1299" height="935" alt="image"
src="https://github.com/user-attachments/assets/2c4d8e47-b5c2-41e2-a0d6-77d48ee3b417"
/>

_Note:_ Many files were impacted by small navigation changes due to the
new tabs, and some were moved.

### New Student Appeals Status Chip

- Create a new status chip status `<status-chip-student-appeal>` to
allow the appeal requests status to be displayed correctly for
"Approved" appeals. Please note, the "decision status" is still
displayed using the `<status-chip-requested-assessment>` that is the
same used for the visualization in the application details and the
appeal read-only page.
- `StudentAppealSharedForm.vue` was converted to have the tabs and to be
shared for the application and student routes.

## E2E Tests

- E2E tests adjusted to include new values returned from the API.
- Added an order to the appeals request to ensure consistent return of
the list. There is no requirement for a particular order, so an order
was chosen.