#5123 - Deleting restrictions: SIMS and SFAS interaction (#5414)
## SFAS to SIMS Bridge

- While importing the restrictions from SFAS, the SQL scripts
responsible for selecting the existing restrictions os SIMS were updated
to ignore deleted records.
- `Bulk-insert-sfas-mapped-restrictions.sql` was updated, even already
having the `is_active = TRUE`, to be explicit about the fact that a
deleted record must be excluded.

## SIMS to SFAS Bridge

- While deleting the restriction on SIMS, the restriction is already
marked as `inactive`, as shown below.

https://github.com/bcgov/SIMS/blob/0e0589951c6b928eabd123bb0eac5fb608cffd03/sources/packages/backend/apps/api/src/services/restriction/student-restriction.service.ts#L323-L332
- Change the select query that defines the records to be sent to SFAS to
also include deleted records.
- Included an `order by updated_at` to ensure records will always be
added to the file in the same order. The record type is still the main
order, and applications are still ordered by intensity, as they were
before. Got a quick approval from @CarlyCotton on this.

### Potential PROD impact

- Deleted restrictions on PROD will not be communicated to SFAS till
this change is deployed. Once deployed, the DB will be checked to ensure
deleted restrictions will be manually updated on SFAS, and expect zero
to a few records only.

## E2E Tests

- Existing E2E tests were adapted to include changed scenarios.

### Minor refactor

- E2E test method `findAndSaveRestriction` was promoted to the factory
to allow it to be shared.