#4968 - Institution Program Restrictions: Suspension - UI - Part 1 (#5430)
## PR Scoppe

- PR intended to be the first part of the UI changes for the ticket.
- Only methods in the frontend/backend related to adding the new
institution restriction were changed.
- Viewing the data, updating the restrictions list, and resolving the
restrictions are not part of this PR.

## Add new restrictions modal

<img width="731" height="529" alt="image"
src="https://github.com/user-attachments/assets/a896b5b0-c5b6-48d7-a6e7-aa89b4da53ce"
/>

<img width="735" height="813" alt="image"
src="https://github.com/user-attachments/assets/958a883b-427d-4330-9fe5-1e1da0a3803c"
/>

### Data load

- Reasons (a.k.a. restrictions descriptions), program list, and
locations list are loaded on demand only when the modal is shown.

#### Locations

- For simplicity, reused the method `allInstitutionLocations` to get all
locations available for an institution. The method does retrieve more
data than required, but this was not considered an issue since it should
return very few records ( 1 to 3 most of the time, maybe 15 for an
exceptional case).

#### Program list:

- Enabled the same endpoint, implemented as
`getProgramsListForInstitutions` for institutions. The method was kept
the same to be "client shared" by the Vue application. If there is a
strong reason to create a different method from the institution's one,
please share the motivation, and we can have it adjusted.
- The current method was already returning approved and active programs,
as per this requirement.
- Enabled only client-side filtering since the amount of data expected
to be loaded and the number of times the feature would be used do not
seem strong enough to execute a server-side search.

#### Restrictions descriptions (a.k.a. reasons)

- Endpoint method `getRestrictionReasons` received a minor refactor to
allow the retrieval of institution restrictions.

### Post New Restriction

- Entity model `institution.model.ts` got new shortcut properties to
allow a single query to retrieve the data to be validated.
- The validation of the existence of an active restriction was treated
as a possible user message to cover a possible race condition where two
users would enter the same data, or if the system DB is under stress.

## E2E Tests

RestrictionAESTController(e2e)-addInstitutionRestriction.
Should throw a bad request exception when no locations were provided.
(99 ms)
√ Should add multiple institution restrictions when a valid payload with
multiple locations IDs is submitted.
√ Should create the institution restriction when there is already an
institution restriction, but it is inactive.
√ Should throw an unprocessable entity exception when an active
institution restriction already exists for at least one location.
√ Should throw a not found exception when the provided institution is
not found.
√ Should throw an unprocessable entity exception when the location does
not exist associated with the institution.
√ Should throw an unprocessable entity exception when the program does
not exist associated with the institution.
√ Should throw an unprocessable entity exception when the restriction ID
does not exist.
√ Should throw a forbidden exception when the user does not have
permission.
√ Should throw a bad request exception when no locations were provided.
√ Should throw a bad request exception when too many locations are
provided.

## Outside PR scope

- Adjusted the modal min width and removed the max/min when in
fullscreen.