#4968 - Institution Program Restrictions: Suspension (UI - add and display restriction) (#5450)
## Restrictions Summary

- Added `Location` and `Programs` columns to the table and enabled
sorting for them.
- Sorting on column `Status` was kept in order to allow it to display
all `Active` items at the top, following the same behavior that was in
place and the same as the students' restrictions.

<img width="1278" height="660" alt="image"
src="https://github.com/user-attachments/assets/268543e8-6edc-4e2b-ae06-292c193f502c"
/>

_Note:_ resolving restrictions is not part of this ticket. The value
added in the above table is a result of the DB manipulation with the
sole purpose of taking the screenshot.

<img width="733" height="523" alt="image"
src="https://github.com/user-attachments/assets/4933670e-c555-48ff-972f-5a76d85127fe"
/>

_Note:_ the only AC related to viewing a restriction is the one that
mentions "Add View Button". No work was actually done on this modal, and
it is here for the sake of demonstration that it is able to display the
data.

## Other Changes

- Included an extra E2E for institution restriction created as
[suggested during previous
PR](https://github.com/bcgov/SIMS/pull/5430#discussion_r2586482905).
- Removed the hardcoded `Designation` institution type present on the
API and student "add restriction".
- Minor refactor in the student restrictions creation modal to remove
errors and adjust the `processing` status (`processing` was added to the
Vue template but not present at the `setup`).

## E2E Tests

RestrictionAESTController(e2e)-getReasonsOptionsList.
- √ Should throw a bad request exception when requesting federal
restrictions reasons.
- √ Should throw a bad request exception when category is not provided.
- Should get reasons options list for a restriction type when the
request is valid.
- √ Should get Provincial restrictions reasons list filtered by specific
category when Provincial restrictions are requested for a category.
- √ Should get Institution restrictions reasons list filtered by
specific category when Institution restrictions are requested for a
category.

RestrictionAESTController(e2e)-addInstitutionRestriction.
- √ Should throw an unprocessable entity exception when trying to add a
provincial restriction.

## Outside scope but open for discussion

- The method `getAllRestrictionCategories` was not refactored beyond the
minimal required to apply the accurate filter. The result is misleading
because the distinct will return the first item of a distinct group to
allow an ID to be returned, but the returned ID has no purpose, and
returning it seems to be a mistake.
- Proposed refactors:
  - Stop returning the ID;
  - Replace the `distinct on` with a regular `distinct`.
  - Pass the restriction type as a parameter.
  - Adjust the API/Web DTOs.