#2994 - T4A - Upload File to Student Account - Part 1 (#5495)
# Overview

- Created two new queues to allow the discovery and processing of all
T4A files available.
_Note:_ Some performance metrics are collected and added to logs with
the intention of supporting SFTP troubleshooting in the future.
- Class `sftp-integration-base.ts` was used to keep the pattern in place
for SFTP-related integrations, but nothing much specialized was required
since the file is downloaded and no content processing is required.
- Adapted the method `downloadResponseFile` to be used to download the
file in an attempt to keep the pattern used by the other SFTP-related
integrations.
- Methods `getClient` and `ensureClientClosed` are now exposed to allow
sharing the SFTP client connection.

## `t4a-upload-enqueuer`

- Scan the T4A IN folder to list all files in all directories(first
level only, non-recursive). At least one folder must be present,
starting with 4 digits that should represent the year, but it is
flexible to allow further content, for instance, but not limited to:
2025, 2025-1, 2025-Reviewed. The folder name became the file group name
to be displayed to the student.
- This scheduler runs every day from Feb to May, but it is expected to
receive a one-time load of 40k to be processed.
- The list of files present in the SFTP should be broken into batches
and queued to the `t4a-upload` to be uploaded to the student account.
- Configuration `maxFileUploadsPerBatch` allows the batch size to be
configured. For a single batch, a single SFTP client connection will be
created.

## `t4a-upload`

- Receives a list of files to be downloaded from the SFTP and uploaded
to the respective student account.
- Receives the relative path to be processed, which will then be
combined with the SFTP T4A folder. This prevents the queue from
accessing any other SFTP folder.
- A unique ID is associated with the file and will be used for logs,
file creation, and S3 upload, preventing the SIN number from being
logged as much as possible.
- Enabled the configuration of the Bull setting `removeOnComplete` to
remove the execution history from Redis once completed, also preventing
the SIN number from being accessible longer than required.

### Log Sample for a Single File

```console
[T4AUploadProcessor] Processing file unique ID 49b1eb7a-4c32-47ff-a823-14347dcd5c0b.
[T4AUploadProcessor] Found student ID 1.
[T4AUploadProcessor] Start download from the SFTP.
[T4AUploadProcessor] File downloaded in 654.71ms.
[T4AUploadProcessor] Start upload to the student account.
[T4AUploadProcessor] Log details
[T4AUploadProcessor] Uploading file 2025-LOAD-TEST-T4A-2025-12-11.pdf to S3 storage.
[T4AUploadProcessor] File 2025-LOAD-TEST-T4A-2025-12-11.pdf uploaded to S3 storage.
[T4AUploadProcessor] Saving the file 2025-LOAD-TEST-T4A-2025-12-11.pdf to database.
[T4AUploadProcessor] Adding the file: 2025-LOAD-TEST-T4A-2025-12-11.pdf to the virus scan queue.
[T4AUploadProcessor] File 2025-LOAD-TEST-T4A-2025-12-11.pdf has been added to the virus scan queue.
[T4AUploadProcessor] Student file ID 2578 created and notification saved.
[T4AUploadProcessor] File created in 69.57ms.
[T4AUploadProcessor] Log details
[T4AUploadProcessor] Processing file unique ID c105bfa4-e122-4416-97d4-0b5a205e6ad7.
[T4AUploadProcessor] Found student ID 1.
[T4AUploadProcessor] Start download from the SFTP.
[T4AUploadProcessor] File downloaded in 268.87ms.
[T4AUploadProcessor] Start upload to the student account.
[T4AUploadProcessor] Log details
[T4AUploadProcessor] Uploading file 2025-LOAD-TEST-T4A-2025-12-11.pdf to S3 storage.
[T4AUploadProcessor] File 2025-LOAD-TEST-T4A-2025-12-11.pdf uploaded to S3 storage.
[T4AUploadProcessor] Saving the file 2025-LOAD-TEST-T4A-2025-12-11.pdf to database.
[T4AUploadProcessor] Adding the file: 2025-LOAD-TEST-T4A-2025-12-11.pdf to the virus scan queue.
[T4AUploadProcessor] File 2025-LOAD-TEST-T4A-2025-12-11.pdf has been added to the virus scan queue.
[T4AUploadProcessor] Student file ID 2579 created and notification saved.
[T4AUploadProcessor] File created in 61.09ms.
```

### `t4a-upload` Queue Payload Sample

```json
{
  "jobData": {
    "referenceDate": "2025-12-10T21:36:55.532Z",
    "files": [
      {
        "relativeFilePath": "2025-LOAD-TEST\\000000300.pdf",
        "uniqueID": "3a99bef6-4838-44ba-bb40-f38568f74e9a"
      },
      {
        "relativeFilePath": "2025-LOAD-TEST\\000000301.pdf",
        "uniqueID": "61c3c294-dd9b-40c4-b0f4-3a380d9b92bb"
      }
    ]
  }
}
```

## API Side Effects

- Method `createFile` was moved to a shared service to allow it to be
used by the API and Integrations lib. - _Note:_ multiple files impacted
with minor changes.
- The core functionality of this method was kept, and there is no plan
to add efforts to change it. Comments related to the previous
implementation may be raised, but will be dealt with outside this PR
review scope.
- Method `getStudentFileByUniqueName` removed because no references were
found.

## E2E Tests Side Effects

- Queue-consumers E2E tests started to fail due to a missing Redis
connectivity (E2E tests are not using Redis), which indicates an issue
with the `QueueModuleMock` during the `overrideImportsMetadata`. To
resolve it, a more updated and out-of-the-box method from Nestjs was
used: `overrideModule`. This method was not in place when the custom
method `overrideImportsMetadata` was created. _Note:_ `ZeebeModule`
requires further work to change to use `overrideModule`. A new effort
will be done in the next PR, or a ticket will be created to have it
changed.

## DB Migration

- Create a single migration to have both queues created.

### DB Migration Rollback

<img width="795" height="160" alt="image"
src="https://github.com/user-attachments/assets/e4419800-90d5-41a4-b801-bf08122643fa"
/>

## TODO

- Create an env variable for T4A IN folder, removing const
`T4A_SFTP_IN_FOLDER`.
- Archiving on success.
- Validate SIN prior to searching the student.
- Further performance benchmarks with different configurations for the
batch sizes.
- Detect SFTP connection issues and interrupt the process to avoid an
attempt on any remaining files.
- E2E tests for the new queues.
- Update the wiki with the new schedulers.

## Ministry UI with Uploaded Files

<img width="1186" height="554" alt="image"
src="https://github.com/user-attachments/assets/0281dcdf-474f-4390-8ef2-bd27aa7e03cf"
/>

## Student UI with Uploaded Files

<img width="1214" height="665" alt="image"
src="https://github.com/user-attachments/assets/def79f34-04e6-4b91-8a90-fc50c27d8602"
/>